requirements.txt
Scrapy
# psycopg2
psycopg2-binary
redis
pymongo
